+++ public/pages/signup.html
@@ -39,6 +39,7 @@
             </label>
         </form>
     </div>
+    <div><p> If You Have An Account<a href="/">Log In</a></p></div>
 <script src="../../public/js/request.js"></script>
 <script src="../../public/js/domSignup.js"></script>
 </body>
